MIX-123: fix cached queries
