+++ packages/supabase/src/queries/index.ts
@@ -1,21 +1,22 @@
-import { unstable_cache } from "next/cache";
-import { createClient } from "../client/server";
 import { Client } from "../types";
 
 export async function getBookmarksQuery(supabase: Client, userId: string) {
   const { data: bookmark_links } = await supabase
     .from("bookmark_link")
-    .select("*")
+    .select()
+    .eq("user_id", userId)
     .throwOnError();
 
   const { data: bookmarks } = await supabase
     .from("bookmarks")
-    .select("*")
+    .select()
+    .eq("user_id", userId)
     .throwOnError();
 
   const { data: collections } = await supabase
     .from("collections")
-    .select("*")
+    .select()
+    .eq("user_id", userId)
     .throwOnError();
 
   return {
@@ -25,21 +26,3 @@ export async function getBookmarksQuery(supabase: Client, userId: string) {
   };
 }
 
-export const getBookmarkData = async () => {
-  const supabase = createClient();
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) return null;
-
-  return unstable_cache(
-    async () => {
-      return getBookmarksQuery(supabase, user.id);
-    },
-    ["bookmarks", user.id],
-    {
-      tags: [`bookmarks_${user.id}`],
-    }
-  );
-};
